refactor(settings): rename to AIProviderSettingsCard and use Field layout for default model
